requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain_huggingface
python-dotenv
fastapi
uvicorn
pydantic
requests
transformers
torch
torchvision
torchaudio
sentence_transformers
langchain_openai 